requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
openai==1.30.1
python-decouple==3.8
tabulate==0.9.0